FlightsTable is finished
